pyproject.toml
[project]
name = "pixel-map"
version = "0.0.1"
description = "Python CLI tool for plotting geo files in the terminal"
authors = [
    {name = "Kamil Raczycki", email = "kraczycki@kraina.ai"},
]
dependencies = [
    "rich>=12.0.0",
    "typer[all]>=0.9.0",
    "img2unicode[n2]>=0.1a11",
    "geopandas>=0.6",
    "numpy>=1.26.0",
    "shapely>=2",
    "contextily>=1.6.2",
    "matplotlib>=3.9.2",
    "pillow>=11.0.0",
    "pyarrow>=16.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
